Split out interpreters for better dependencies
